# api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "wwg API",
    "description": "The Wasatch Wargamers Campaign App API.",
    "version": "v1"
  },
  "paths": {
...
    "/api/auth/masquerade/end": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Ends a masquerade (decision 0012): back to the Admin's own session, without a password, if\nthe masquerade may still go on. 409 if the session isn't a masquerade.",
        "operationId": "EndMasquerade",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/api/armies/{id}/turns": {
      "get": {
        "tags": [
          "Turns"
        ],
        "summary": "An army's turns, newest first, with their orders and history (its commander, the Umpire and\nAdmins: the visibility rule).",
        "operationId": "ListArmyTurns",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArmyTurnDetails"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/campaigns/{id}/turns": {
      "post": {
        "tags": [
          "Turns"
        ],
        "summary": "Starts the next turn (Umpire or Admin), once every army's turn in the open one is Completed\nand every unit is on the map: the open turn closes and the next opens, a Draft for every\narmy. Every commander is emailed.",
        "operationId": "StartNextTurn",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignTurnsResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "get": {
        "tags": [
          "Turns"
        ],
        "summary": "The campaign's turns and where each stands (every member). The Umpire sees every army's\nturn, and what stops the campaign (or, once it's running, the next turn) starting; a\ncommander sees their own army's, and everyone the counts.",
        "operationId": "ListTurns",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignTurnsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/api/army-turns/{id}/orders/{unitId}": {
      "put": {
        "tags": [
          "Turns"
        ],
        "summary": "Gives a unit its order for the turn: Move along a path of adjacent hexes from its current\none, inside the grid (and for the commander, one its movement class can afford this turn:\nDESIGN.md §5.2); or Hold, where it is. Replaces any order it had. The army's commander gives orders while the turn is a\nDraft in the open campaign turn; the Umpire and Admins while it's a Draft or Submitted, on\nthe commander's behalf (decision 0011), which goes in the turn's history.",
        "operationId": "GiveOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "unitId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GiveOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitPosition"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Turns"
        ],
        "summary": "Takes back a unit's order for the turn: the army's commander while it's a Draft; the\nUmpire and Admins while it's a Draft or Submitted, which goes in the turn's history.",
        "operationId": "UndoOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "unitId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArmyTurnDetails": {
        "required": [
          "id",
          "turn",
          "open",
          "status",
          "submittedAt",
          "completedAt",
          "orders",
          "history"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The army turn's ID (orders are given through it).",
            "format": "uuid"
          },
          "turn": {
            "type": "integer",
            "description": "The campaign turn's number.",
            "format": "int32"
          },
          "open": {
            "type": "boolean",
            "description": "Whether it's in the open campaign turn."
          },
          "status": {
            "description": "Draft, Submitted or Completed.",
            "$ref": "#/components/schemas/ArmyTurnStatus"
          },
          "submittedAt": {
            "type": [
              "null",
              "string"
            ],
            "description": "When it was last submitted (UTC).",
            "format": "date-time"
          },
          "completedAt": {
            "type": [
              "null",
              "string"
            ],
            "description": "When it was approved (UTC).",
            "format": "date-time"
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UnitPosition"
            },
            "description": "Its units' orders, and so positions after it."
          },
          "history": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArmyTurnEventDto"
            },
            "description": "What happened to it, oldest first."
          }
        },
        "description": "An army's turn, with its orders and history."
      },
...
      "ArmyTurnStatus": {
        "enum": [
          "Draft",
          "Submitted",
          "Completed"
        ],
        "description": "Where an army's turn is."
      },
...
      "CampaignStage": {
        "enum": [
          "Setup",
          "Running"
        ],
        "description": "Where a campaign is: setting up (turn 0), or running (turn 1 or later is open)."
      },
...
      "CampaignTurnsResponse": {
        "required": [
          "stage",
          "openTurn",
          "turns",
          "startProblems"
        ],
        "type": "object",
        "properties": {
          "stage": {
            "description": "Setting up (turn 0) or running.",
            "$ref": "#/components/schemas/CampaignStage"
          },
          "openTurn": {
            "type": "integer",
            "description": "The open turn's number (0 while setting up).",
            "format": "int32"
          },
          "turns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CampaignTurnSummary"
            },
            "description": "Every turn so far, oldest first."
          },
          "startProblems": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "For the Umpire: what stops the campaign starting (while setting up) or the next turn (once running)."
          }
        },
        "description": "Where the campaign's turns are."
      },
...
      "GiveOrderRequest": {
        "required": [
          "kind",
          "path"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "description": "Move (along the path) or Hold (stay where it is).",
            "$ref": "#/components/schemas/OrderKind"
          },
          "path": {
            "maxItems": 50,
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/Hex"
            },
            "description": "A Move's steps: the hexes it passes through in order, from next to the unit's hex to where it\nends. Adjacent, inside the grid, and within what the unit can move in a turn."
          }
        },
        "description": "A unit's order for the turn."
      },
...
      "Hex": {
        "required": [
          "q",
          "r"
        ],
        "type": "object",
        "properties": {
          "q": {
            "type": "integer",
            "format": "int32"
          },
          "r": {
            "type": "integer",
            "format": "int32"
          }
        },
        "description": "A hex in a campaign's grid, by its axial coordinates (q east, r south-east)."
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "OrderKind": {
        "enum": [
          "Move",
          "Hold"
        ],
        "description": "What a unit does in a turn."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "TokenResponse": {
        "required": [
          "accessToken",
          "expiresIn"
        ],
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "The bearer token for API calls."
          },
          "expiresIn": {
            "type": "integer",
            "description": "Seconds until the access token expires; refresh before then.",
            "format": "int32"
          }
        },
        "description": "A signed-in session. The access token goes in `Authorization: Bearer …`; the refresh\ntoken is never in a body: it's set as an HttpOnly cookie that only /api/auth receives."
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "The accessToken from register, login or refresh.",
        "scheme": "bearer"
      }
    }
  },
...
}